requirements_freeze.txt
absl-py==0.12.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astroid==2.6.6
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
audioread==2.1.9
autopep8==1.5.7
azure-functions==1.7.2
backcall==0.2.0
base58==2.1.0
bidict==0.21.2
bleach==3.3.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.3.2
cursor==1.3.4
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
docopt==0.6.2
ecdsa==0.16.1
entrypoints==0.3
ffmpeg-python==0.2.0
flake8==3.9.2
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
gpg==1.16.0
greenlet==1.1.2
grpcio==1.34.1
h5py==3.1.0
halo==0.0.28
harview==0.2.2
hdwallet==1.1.1
idna==2.10
imbalanced-learn==0.8.0
imblearn==0.0
ipaddress==1.0.23
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-http-over-ws==0.0.8
jupyterlab-pygments==0.1.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
librosa==0.8.1
llvmlite==0.36.0
log-symbols==0.0.14
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mnemonic==0.19
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
neovim==0.3.1
nest-asyncio==1.5.1
notebook==6.4.0
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip-chill==1.0.1
pluggy==1.0.0
pooch==1.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.3
ptyprocess==0.7.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.7.0
pycparser==2.20
pycryptodomex==3.10.1
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.9.0
pylint==2.9.6
pynvim==0.4.3
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.17.3
pysha3==1.0.2
python-dateutil==2.8.1
python-engineio==4.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-socketio==5.4.0
python-xlib==0.31
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
requirements-parser==0.2.0
resampy==0.2.2
rope==0.21.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
sentry-sdk==1.1.0
six==1.15.0
sklearn==0.0
smmap==4.0.0
snowballstemmer==2.1.0
SoundFile==0.10.3.post1
spell==0.41.77
spinners==0.0.24
starlette==0.14.2
statsmodels==0.12.2
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
tree-sitter==0.19.0
typing-extensions==3.7.4.3
ujson==4.2.0
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.0
Werkzeug==2.0.1
wrapt==1.12.1
yapf==0.31.0
yellowbrick==1.3.post1
